# swagger.yml
basePath: /
consumes:
- application/json
definitions:
  Comment:
    properties:
      body:
        description: Body of the comment
        maxLength: 255
        minLength: 5
        type: string
        x-go-name: Body
      createdAt:
        description: Time comment was created at
        type: string
        x-go-name: CreatedAt
      feedbackId:
        description: Id of the feedback this comment is related to
        format: int64
        minimum: 1
        type: integer
        x-go-name: FeedbackId
      id:
        description: Id of the comment
        minimum: 1
        type: string
        x-go-name: Id
      updatedAt:
        description: Time comment was updated at
        type: string
        x-go-name: UpdatedAt
      userId:
        description: Id of the user who created the comment
        format: int64
        minimum: 1
        type: integer
        x-go-name: UserId
    type: object
    x-go-package: product-feedback/comment
  Feedback:
    properties:
      createdAt:
        description: Time this vote was created at
        type: string
        x-go-name: CreatedAt
      feedbackId:
        description: Feedback id that this vote is related to
        format: int64
        minimum: 1
        type: integer
        x-go-name: FeedbackId
      id:
        description: Id of the vote
        format: int64
        minimum: 1
        type: integer
        x-go-name: Id
      updatedAt:
        description: Time this vote was updated at
        type: string
        x-go-name: UpdatedAt
      userId:
        description: Id of a user that created this vote
        format: int64
        minimum: 1
        type: integer
        x-go-name: UserId
    type: object
    x-go-name: Vote
    x-go-package: product-feedback/vote
  User:
    properties:
      avatar_url:
        description: Avatar URL
        type: string
        x-go-name: AvatarUrl
      created_at:
        description: Time the user was created at
        type: string
        x-go-name: CreatedAt
      email:
        description: Email of the user
        example: user@company.com
        type: string
        x-go-name: Email
      id:
        description: Id of the user
        format: int64
        minimum: 1
        type: integer
        x-go-name: Id
      name:
        description: Name of the user
        maxLength: 50
        minLength: 2
        type: string
        x-go-name: Name
      updated_at:
        description: Time the user was updated at
        type: string
        x-go-name: UpdatedAt
      user_name:
        description: User name
        maxLength: 50
        minLength: 2
        type: string
        x-go-name: UserName
    type: object
    x-go-package: product-feedback/user
info:
  description: Documentation Product Feedback API
  title: Product Feedback API
  version: 1.0.0
paths:
  /api/auth/github:
    get:
      description: Redirects to GitHub authentication
      operationId: loginWithGitHub
      responses:
        "302":
          $ref: '#/responses/foundResponse'
      tags:
      - auth
  /api/auth/github/callback:
    get:
      description: Redirects to GitHub authentication
      operationId: loginWithGitHub
      responses:
        "302":
          $ref: '#/responses/foundResponse'
        "403":
          $ref: '#/responses/errorResponse'
      tags:
      - auth
  /api/auth/refresh-token:
    post:
      description: Redirects to GitHub authentication
      operationId: refreshAccessToken
      responses:
        "200":
          $ref: '#/responses/refreshAccessTokenResponse'
        "403":
          $ref: '#/responses/errorResponse'
      tags:
      - auth
  /api/comments:
    get:
      description: Returns a list of all comments in the system
      operationId: getAllComments
      responses:
        "200":
          $ref: '#/responses/getAllCommentsResponse'
      security:
      - Bearer: []
      tags:
      - comments
    post:
      description: Creates a comment
      operationId: createComment
      responses:
        "200":
          $ref: '#/responses/createCommentResponse'
      security:
      - Bearer: []
      tags:
      - comments
  /api/comments/:id:
    get:
      description: Returns comment by id
      operationId: getCommentById
      responses:
        "200":
          $ref: '#/responses/okResponse'
        "404":
          $ref: '#/responses/errorResponse'
      security:
      - Bearer: []
      tags:
      - comments
  /api/feedback:
    get:
      description: Returns all product feedback in the system
      operationId: getAllFeedback
      responses:
        "200":
          $ref: '#/responses/getAllFeedbackResponse'
      security:
      - Bearer: []
      tags:
      - feedback
    post:
      description: Create product feedback
      operationId: createFeedback
      responses:
        "200":
          $ref: '#/responses/createFeedbackResponse'
      security:
      - Bearer: []
      tags:
      - feedback
  /api/feedback/:id:
    delete:
      description: Delete product feedback
      operationId: deleteFeedback
      responses:
        "200":
          $ref: '#/responses/okResponse'
        "404":
          $ref: '#/responses/errorResponse'
      security:
      - Bearer: []
      tags:
      - feedback
    get:
      description: Returns product feedback by id
      operationId: getFeedbackById
      responses:
        "200":
          $ref: '#/responses/getFeedbackByIdResponse'
        "404":
          $ref: '#/responses/errorResponse'
      security:
      - Bearer: []
      tags:
      - feedback
    put:
      description: Returns product feedback by id
      operationId: updateFeedback
      responses:
        "200":
          $ref: '#/responses/okResponse'
        "400":
          $ref: '#/responses/errorResponse'
        "404":
          $ref: '#/responses/errorResponse'
      security:
      - Bearer: []
      tags:
      - feedback
  /api/users/:id:
    get:
      description: Returs user data by id
      operationId: getUserById
      responses:
        "200":
          $ref: '#/responses/getUserResponse'
        "404":
          $ref: '#/responses/errorResponse'
      security:
      - Bearer: []
      tags:
      - users
  /api/users/me:
    get:
      description: Returs user data
      operationId: getMe
      responses:
        "200":
          $ref: '#/responses/getUserResponse'
      security:
      - Bearer: []
      tags:
      - users
  /api/votes:
    get:
      description: Returns a list of votes in the system
      operationId: getAllVotes
      responses:
        "200":
          $ref: '#/responses/getAllVotesResponse'
      security:
      - Bearer: []
      tags:
      - votes
    post:
      description: Creates a vote
      operationId: createVote
      responses:
        "200":
          $ref: '#/responses/createVoteResponse'
      security:
      - Bearer: []
      tags:
      - votes
  /api/votes/:id:
    delete:
      description: Deletes a vote
      operationId: deleteVote
      responses:
        "200":
          $ref: '#/responses/okResponse'
        "404":
          $ref: '#/responses/errorResponse'
      security:
      - Bearer: []
      tags:
      - votes
produces:
- application/json
responses:
  createCommentResponse:
    description: Create comment response
    schema:
      properties:
        commentId:
          description: id of a newly created comment
          type: string
          x-go-name: CommentId
      type: object
  createFeedbackResponse:
    description: OK response
    schema:
      properties:
        feedbackId:
          description: id of a newly created feedback
          type: string
          x-go-name: FeedbacId
      type: object
  createVoteResponse:
    description: OK response
    schema:
      properties:
        voteId:
          description: id of a newly created vote
          type: string
          x-go-name: VoteId
      type: object
  errorResponse:
    description: Error response
    schema:
      properties:
        message:
          description: Error message
          type: string
          x-go-name: Message
      type: object
  foundResponse:
    description: Redirects to login route
  getAllCommentsResponse:
    description: Returns a list of all comments in the system
    schema:
      items:
        $ref: '#/definitions/Comment'
      type: array
  getAllFeedbackResponse:
    description: Returns all product feedback in the system
    schema:
      items:
        $ref: '#/definitions/Feedback'
      type: array
  getAllVotesResponse:
    description: OK response
    schema:
      items:
        $ref: '#/definitions/Feedback'
      type: array
  getFeedbackByIdResponse:
    description: Returns product feedback by id
    schema:
      $ref: '#/definitions/Feedback'
  getUserResponse:
    description: OK response
    schema:
      $ref: '#/definitions/User'
  okResponse:
    description: OK response
    schema:
      properties:
        message:
          description: OK message
          type: string
          x-go-name: Message
      type: object
  refreshAccessTokenResponse:
    description: OK response
    schema:
      properties:
        accessToken:
          description: Access token
          type: string
          x-go-name: AccessToken
      type: object
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"